# server/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Uprep API
  description: |
    ## Structure

    ### The Envelope
    Every response is contained by an envelope. That is, each response has a
    predictable set of keys with which you can expect to interact:
    ```json
    {
        "meta": {
            "code": 200
        },
        "data": {
            ...
        },
        "pagination": {
            "next_url": "...",
            "next_max_id": "13872296"
        }
    }
    ```

    #### META
    The meta key is used to communicate extra information about the response to
    the developer. If all goes well, you'll only ever see a code key with value
    200. However, sometimes things go wrong, and in that case you might see a
    response like:
    ```json
    {
        "meta": {
            "error_type": "OAuthException",
            "code": 400,
            "error_message": "..."
        }
    }
    ```

    #### DATA
    The data key is the meat of the response. It may be a list or dictionary,
    but either way this is where you'll find the data you requested.
    #### PAGINATION
    Sometimes you just can't get enough. For this reason, we've provided a
    convenient way to access more data in any request for sequential data.
    Simply call the url in the next_url parameter and we'll respond with the
    next set of data.
    ```json
    {
        ...
        "pagination": {
...
            "next_max_id": "13872296"
        }
    }
    ```
...


...
host: dev-uprep.nijel.org
...
basePath: /api/
...
schemes:
...
consumes:
...
produces:
...
paths:
...
  /grades:
    get:
      tags:
      - "grades"
      summary: "Gets list of available grades"
      description: "Gets the full list of available grades which can be filtered using the school id"
      operationId: "getGrades"
      produces:
      - "application/json"
      parameters:
      - name: schoolId
        in: query
        description: ID of school
        type: integer
        required: true
      responses:
        200:
          description: OK
          schema:
            type: object
            properties:
              data:
                type: "array"
                items:
                  $ref: "#/definitions/Grade"
...
definitions:
...
  Teacher:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
      subjects:
        type: "array"
        items:
          $ref: "#/definitions/Subject"
  Subject:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
  Grade:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
      teachers:
        type: "array"
        items:
          $ref: "#/definitions/Teacher"
...